Move getProducts to /data/coinbase. Add to menu model only on success
